Align tensor primitive dispatch with runtime
